+++ crates/action-graph/src/action_graph_builder.rs
@@ -20,7 +20,6 @@ use moon_task_args::parse_task_args;
 use moon_workspace_graph::{GraphConnections, WorkspaceGraph, tasks::TaskGraphError};
 use petgraph::prelude::*;
 use rustc_hash::{FxHashMap, FxHashSet};
-use starbase_utils::glob::GlobSet;
 use std::mem;
 use std::sync::Arc;
 use tracing::{debug, instrument, trace};
@@ -352,7 +351,7 @@ impl<'query> ActionGraphBuilder<'query> {
             .await?;
 
         // Only insert this action if a root was located
-        if let Some(root) = output.root {
+        if let Some(root) = &output.root {
             let abs_root = toolchain.from_virtual_path(root.any_path());
             let rel_root = abs_root
                 .relative_to(&self.app_context.workspace_root)
@@ -360,14 +359,10 @@ impl<'query> ActionGraphBuilder<'query> {
 
             // Determine if we're in the dependencies workspace
             let in_project = project.root == abs_root;
-            let in_workspace = if let Some(globs) = output.members {
-                if in_project {
-                    true // Root always in the workspace
-                } else {
-                    GlobSet::new(&globs)?.matches(project.source.as_str())
-                }
+            let in_workspace = if in_project {
+                true // Root always in the workspace
             } else {
-                true
+                toolchain.in_dependencies_workspace(&output, project.source.as_str())?
             };
 
             // If not in the dependencies workspace (if there is one),