+++ crates/actions/src/actions/sync_workspace.rs
@@ -102,12 +102,16 @@ pub async fn sync_workspace(
     if toolchain_registry.has_plugins() {
         debug!("Syncing operations from toolchains");
 
+        let app_context = Arc::clone(&app_context);
+
         operation_futures.push(task::spawn(async move {
             let mut ops = vec![];
 
             for sync_result in toolchain_registry
-                .sync_workspace_all(|registry, _| SyncWorkspaceInput {
+                .sync_workspace_all(|registry, toolchain| SyncWorkspaceInput {
                     context: registry.create_context(),
+                    toolchain_config: registry
+                        .create_config(&toolchain.id, &app_context.toolchain_config),
                 })
                 .await?
             {