+++ crates/app/src/commands/docker/prune.rs
@@ -1,71 +1,151 @@
 use super::{DockerManifest, MANIFEST_NAME, docker_error::AppDockerError};
 use crate::session::MoonSession;
+use moon_actions::operations::exec_plugin_command;
 use moon_bun_tool::BunTool;
 use moon_common::Id;
 use moon_config::PlatformType;
 use moon_deno_tool::DenoTool;
 use moon_node_lang::PackageJsonCache;
 use moon_node_tool::NodeTool;
-use moon_pdk_api::PruneDockerInput;
+use moon_pdk_api::{InstallDependenciesInput, LocateDependenciesRootInput, PruneDockerInput};
 use moon_platform::PlatformManager;
+use moon_project::Project;
 use moon_rust_tool::RustTool;
 use moon_tool::DependencyManager;
-use moon_toolchain_plugin::{ToolchainPlugin, ToolchainRegistry};
 use rustc_hash::FxHashSet;
 use starbase::AppResult;
 use starbase_utils::{fs, json};
+use std::path::PathBuf;
+use std::sync::Arc;
+use tokio::task::JoinSet;
 use tracing::{debug, instrument};
 
+#[derive(Default)]
+struct PruneToolchainInstance {
+    deps_root: PathBuf,
+    projects: Vec<Arc<Project>>,
+    toolchain_id: Id,
+}
+
 #[instrument(skip_all)]
-pub async fn prune_toolchain(
-    session: &MoonSession,
-    toolchain_registry: &ToolchainRegistry,
-    toolchain: &ToolchainPlugin,
-) -> AppResult {
+pub async fn prune_toolchains(session: &MoonSession, manifest: &DockerManifest) -> AppResult {
     let project_graph = session.get_project_graph().await?;
+    let toolchain_registry = session.get_toolchain_registry().await?;
 
-    if session
-        .workspace_config
-        .docker
-        .prune
-        .delete_vendor_directories
-    {
-        if let (Some(vendor_name), Some(manifest_name)) = (
-            &toolchain.metadata.vendor_dir_name,
-            &toolchain.metadata.manifest_file_name,
-        ) {
-            debug!(
-                "Removing {} vendor directories ({})",
-                toolchain.metadata.name, vendor_name
-            );
+    // Collect all dependency roots and which projects belong to it
+    let mut deps_roots: Vec<PruneToolchainInstance> = vec![];
 
-            fs::remove_dir_all(session.workspace_root.join(vendor_name))?;
+    for project_id in &manifest.focused_projects {
+        let project = project_graph.get(project_id)?;
 
-            for source in project_graph.sources().values() {
-                let project_root = source.to_logical_path(&session.workspace_root);
+        for locate_result in toolchain_registry
+            .locate_dependencies_root_many(
+                project.get_enabled_toolchains(),
+                |registry, toolchain| LocateDependenciesRootInput {
+                    context: registry.create_context(),
+                    starting_dir: toolchain.to_virtual_path(&project.root),
+                },
+            )
+            .await?
+        {
+            if let Some(root) = locate_result
+                .output
+                .root
+                .as_ref()
+                .and_then(|root| root.real_path())
+            {
+                let toolchain = toolchain_registry.load(&locate_result.id).await?;
 
-                // Only remove if there's a sibling manifest
-                if project_root.join(manifest_name).exists() {
-                    fs::remove_dir_all(project_root.join(vendor_name))?;
+                if project.root != root
+                    && !toolchain
+                        .in_dependencies_workspace(&locate_result.output, project.source.as_str())?
+                {
+                    continue;
                 }
+
+                match deps_roots.iter_mut().find(|inst| {
+                    inst.deps_root == root
+                        && inst.toolchain_id.as_str() == locate_result.id.as_str()
+                }) {
+                    Some(entry) => {
+                        entry.projects.push(project.clone());
+                    }
+                    None => {
+                        deps_roots.push(PruneToolchainInstance {
+                            deps_root: root,
+                            projects: vec![project.clone()],
+                            toolchain_id: Id::raw(locate_result.id),
+                        });
+                    }
+                };
             }
         }
     }
 
-    if session.workspace_config.docker.prune.install_toolchain_deps {
-        // TODO
+    if deps_roots.is_empty() {
+        return Ok(None);
     }
 
-    if toolchain.has_func("prune_docker").await {
-        toolchain
-            .call_func_without_output(
-                "prune_docker",
-                PruneDockerInput {
-                    context: toolchain_registry.create_context(),
-                    docker_config: session.workspace_config.docker.prune.clone(),
-                },
-            )
-            .await?;
+    // Then prune and install dependencies for each root (and its projects)
+    let mut set = JoinSet::new();
+
+    for instance in deps_roots {
+        let toolchain_registry = Arc::clone(&toolchain_registry);
+        let toolchain = toolchain_registry.load(instance.toolchain_id).await?;
+        let docker_config = session.workspace_config.docker.prune.clone();
+
+        set.spawn(async move {
+            // Run prune first, so this can remove all development artifacts
+            if toolchain.has_func("prune_docker").await {
+                let _output = toolchain
+                    .prune_docker(PruneDockerInput {
+                        context: toolchain_registry.create_context(),
+                        docker_config: docker_config.clone(),
+                        projects: instance
+                            .projects
+                            .iter()
+                            .map(|project| project.to_fragment())
+                            .collect(),
+                        root: toolchain.to_virtual_path(&instance.deps_root),
+                    })
+                    .await?;
+            }
+
+            // Then run install, so this can only install production dependencies
+            if toolchain.has_func("install_dependencies").await
+                && docker_config.install_toolchain_deps
+            {
+                let output = toolchain
+                    .install_dependencies(InstallDependenciesInput {
+                        context: toolchain_registry.create_context(),
+                        dependency_names: Some(
+                            instance
+                                .projects
+                                .iter()
+                                .flat_map(|project| project.alias.clone())
+                                .collect(),
+                        ),
+                        production_only: true,
+                        root: toolchain.to_virtual_path(&instance.deps_root),
+                    })
+                    .await?;
+
+                if let Some(mut install_command) = output.install_command {
+                    // Always stream output to the console
+                    install_command.stream = true;
+
+                    // Ensure it runs in the dependency root
+                    if install_command.working_dir.is_none() {
+                        install_command.working_dir =
+                            Some(toolchain.to_virtual_path(&instance.deps_root));
+                    }
+
+                    exec_plugin_command(&install_command).await?;
+                }
+            }
+
+            Ok::<_, miette::Report>(())
+        });
     }
 
     Ok(None)
@@ -238,8 +318,6 @@ pub async fn prune(session: MoonSession) -> AppResult {
     }
 
     // Do this later so we only run once for each platform instead of per project
-    let toolchain_registry = session.get_toolchain_registry().await?;
-
     for toolchain_id in toolchains {
         if toolchain_id == "unknown" {
             // Will crash with "Platform unknown has not been enabled"
@@ -298,11 +376,9 @@ pub async fn prune(session: MoonSession) -> AppResult {
                 _ => {}
             };
         }
-
-        if let Ok(toolchain) = toolchain_registry.load(toolchain_id).await {
-            prune_toolchain(&session, &toolchain_registry, &toolchain).await?;
-        }
     }
+
+    prune_toolchains(&session, &manifest).await?;
 
     Ok(None)
 }